refactor: 쓸모없는 클래스를 지우고 통합
